+++ packages/expo-media-library/android/src/main/java/expo/modules/medialibrary/albums/migration/MigrateAlbum.kt
@@ -1,16 +1,15 @@
 package expo.modules.medialibrary.albums.migration
 
-import expo.modules.medialibrary.MediaLibraryUtils.AssetFile
-import android.os.AsyncTask
-import android.content.ContentValues
-import android.provider.MediaStore
 import android.content.ContentUris
+import android.content.ContentValues
 import android.content.Context
 import android.os.Build
+import android.provider.MediaStore
 import androidx.annotation.RequiresApi
-import expo.modules.core.Promise
-import expo.modules.medialibrary.ERROR_UNABLE_TO_MIGRATE
+import expo.modules.kotlin.Promise
+import expo.modules.medialibrary.AlbumException
 import expo.modules.medialibrary.MediaLibraryUtils
+import expo.modules.medialibrary.MediaLibraryUtils.AssetFile
 import java.io.File
 
 @RequiresApi(Build.VERSION_CODES.R)
@@ -19,8 +18,8 @@ internal class MigrateAlbum(
   private val assetFiles: List<AssetFile>,
   private val albumDirName: String,
   private val promise: Promise
-) : AsyncTask<Void?, Void?, Void?>() {
-  override fun doInBackground(vararg voids: Void?): Void? {
+) {
+  fun execute() {
     // Previously, users were able to save different assets type in the same directory.
     // But now, it's not always possible.
     // If album contains movies or pictures, we can move it to Environment.DIRECTORY_PICTURES.
@@ -29,8 +28,7 @@ internal class MigrateAlbum(
       .map { MediaLibraryUtils.getRelativePathForAssetType(it.mimeType, false) }
       .toSet()
     if (assetsRelativePaths.size > 1) {
-      promise.reject(ERROR_UNABLE_TO_MIGRATE, "The album contains incompatible file types.")
-      return null
+      throw AlbumException("The album contains incompatible file types.")
     }
 
     val relativePath = assetsRelativePaths.iterator().next() + File.separator + albumDirName
@@ -50,6 +48,5 @@ internal class MigrateAlbum(
         )
     }
     promise.resolve(null)
-    return null
   }
 }